fix(fast_io): drop unnecessary i32 cast in POLLOUT bitand (#1874)
`revents` is i16 and `libc::POLLOUT` is `c_short` (i16) on Linux, so the
cast to i32 produced a `BitAnd<i16>` mismatch. Both sides are already
i16; remove the cast.

+++ crates/fast_io/tests/io_uring_shared_ring.rs
@@ -124,7 +124,7 @@ mod linux_only {
                 SharedCompletion::PollWrite { op_id, revents } => {
                     assert_eq!(op_id, 202, "poll tag must carry the submitted op_id");
                     assert!(
-                        (revents as i32) & libc::POLLOUT != 0,
+                        revents & libc::POLLOUT != 0,
                         "expected POLLOUT in revents, got {revents:#x}"
                     );
                     saw_poll = true;
